Fix : ClovaApi 관련코드 clovaModule로 리팩토링
