--- src/main/java/com/hwachang/hwachangapi/domain/consultingRoomModule/controller/FileUploadController.java
@@ -1,31 +1,0 @@
-package com.hwachang.hwachangapi.domain.consultingRoomModule.controller;
-
-
-import com.hwachang.hwachangapi.domain.consultingRoomModule.service.NcloudStorageService;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
-import org.springframework.web.multipart.MultipartFile;
-
-import java.io.IOException;
-
-@RestController
-@RequestMapping("/api/storage")
-public class FileUploadController {
-
-    private final NcloudStorageService storageService;
-
-    public FileUploadController(NcloudStorageService storageService) {
-        this.storageService = storageService;
-    }
-
-    @PostMapping("/upload")
-    public ResponseEntity<String> uploadFile(@RequestParam("file") MultipartFile file) {
-        try {
-            String fileUrl = storageService.uploadFile(file);
-            return ResponseEntity.ok("File uploaded successfully: " + fileUrl);
-        } catch (IOException e) {
-            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body("File upload failed: " + e.getMessage());
-        }
-    }
-}
